% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TFA2D_wrapper.R
\name{balance_sample_purity_bandwidth_tf}
\alias{balance_sample_purity_bandwidth_tf}
\title{main EB function}
\usage{
balance_sample_purity_bandwidth_tf(
  d1_data,
  d2_data,
  s1_col_name,
  s2_col_name,
  nna_cutoff,
  permute_time,
  working_dir,
  compare_name,
  bandwidth_factor = 1,
  adjust_purity_flag,
  annot_file
)
}
\arguments{
\item{d1_data}{dataframe for the data of the first dimension, rows are genes, columns are patients in 2 conditions}

\item{d2_data}{dataframe for the data of the second dimension, rows are genes, columns are patients in 2 conditions}

\item{s1_col_name}{names of the columns in the first condition}

\item{s2_col_name}{names of the columns in the second condition}

\item{nna_cutoff}{genes with values in at least the number of samples are included in comparison}

\item{permute_time}{number of permutations}

\item{working_dir}{the directory output files will be deposited in}

\item{compare_name}{a label for the anlaysis}

\item{bandwidth_factor}{a numeric value for smoothing the empirical distribution, the larger the smoother, default to 1}

\item{adjust_purity_flag}{a boolean value if set to True tumor purity is adjusted and the annot_file must be provided}

\item{annot_file}{absolute directory to the sample annotation file, column name must be purity, ID column name must be caseID}
}
\description{
calculate posterior probability of each relationship being significant
}
\examples{
balanced sampling 
}
\keyword{2D}
\keyword{comparison}
\keyword{unpaired}
