% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TFA2D_wrapper.R
\name{tf_map}
\alias{tf_map}
\title{transcriptionFactor_substrate mapping function}
\usage{
tf_map(protData_filename, rnaData_filename, tfNetwork_filename, tf_outputName)
}
\arguments{
\item{protData_filename}{absolute directory to the file of the proteome data, in the format of the tidied up FragPipe TMT-I output}

\item{rnaData_filename}{absolute directory to the file of RNA data, ID is the column name for gene symbol, in the format of the tidied up FragPipe TMT-I output, normalized or not normalized}

\item{tfNetwork_filename}{absolute directory to the TF substrate network file, as an example, merge_ks_omni.tsv.}

\item{tf_outputName}{the file name of the output file of this function}
}
\description{
map the knase and substrate information
}
